ui changes
31/10/2022
"1. About Partnership : on page bottom 'View Partnership benefits' button click should take to top of the page
2. The wisdom exercise carousal screen text are overlapping on mobile ..
3. Ui changes
a. The two paragraphs need to be aligned on the left and right. Currently the lower paragraph seems to be insetted from both side
b. The font size of the question is not consistent with the other screens
c. Even this font size is smaller than other screens. Reflection screen: when you check 'Share this note...' the opacity of the white background should change from 50% to 75%
d. The header icons are getting displaced as we progress from one screen to next (video provided)
e. Header icons need to be spaced out better as per the design (image attached)"

01/11/2022
"1. Search dropdown tried to design as WhatsApp but couldn’t complete - will need some time
2. Wisdom exercise
a. Images updated
b. Days updated on index
c. Routes added
3. While loading body background colour updated"

02/11/2022
"1. the annual button on scroll  overlaps over the add-to- cart button on add to cart page
2. Stress module - Reflection questions font size to be increased and kept consistent with the rest of the screens"

03/11/2022
"1. In-module screens
a. Please check the share icon - it must be an exact square (height and width of the icon need to be the same)
b. The text at the top needs to be the same size as ""1. Heart Disease""
c. Module landing page > Contents list > The audio item title and the icon need to be aligned with the rest of the list (attaching sample screenshot)
2. search text box - search Listing styling to be done - made full width"

04/11/2022
"1. Pencil icon overlapping on user email in view cart page(in app/staging/local).
2. Notification icon not visible for unsubscribe user in mobile view.
3. Guided Audio Meditation Module Index Page changes:  - images changed on dashboard as well as on the toc page "

05/11/2022
"1. landing adverts pages - routing of hamburger menu and Signup for newsletter changes
2. intro screen : 4th screen not as per figma ."