ui changes
12/11/2022
1. Black scroll issue wisdom exercise
2. Change the content of index page of wisdom exercise.its in different language.
3. Wisdom exercise 
a. session 6 - 6th day incomplete. 
b. In 10 session 5day in audio file added multiple audios.6Day,7Day,8Day is incomplete 
c. In 11 session the 3Day, 7Day,8Day,9Day  is incomplete  
d. In 12 session 6Day in audio file extra audio file added.which is not is docs.And 7Day is incomplete. 

14/11/2022
1. Wisdom exercise - audio + reflection - session 10, 11 & 12