ui changes
1. Country List styling
2. the country after selection also showing  in placeholder style
3. vertical spacing between textboxes and button is not very uniform
4. and the width of isd code textbox should be more (atleast four digit character width +227)
5. in noticing thoughts module of adults and teenagers toc button missing on some screen , need to add
6. Adults Self Interest Toc page, text flowing out of main frame
7. Wisdom shorts tile image missing-6 steps with living with compassion-wisdom_shorts_78.svg
8. try-free-and-subscribe page on changing focus to monthly the annual tile looks broken
9. Width Issue in redeem-subscription
10. Website home page Changes
a. Top Title to change to : Learn about yourself, Grow as a person, Flourish in life
b. Top Short description on page to change to:A life-companion to help you overcome stress and anxiety, have healthy relationships, be happier and succeed in life
c. Hamburger menu on Home: need to change to HappierMe for business