ui changes
1. set password page is incomplete
2. Website design change
a. the buttons are not same in length
b. The image of bubbles white space to be removed https://humanwisdoms3.s3.eu-west-2.amazonaws.com/assets/svgs/v1_3/bubbles.svg
c. Website Hamburger Menu Styling to be changed
d. Gift of happiness to be redone after Nidhi's confirmed design
e. video Tour to be added in the home page of website
f. Sign up for newsletter liner to change for website and keep as mentioned in gift landing page
g. Website subscription functionalities code need to be readded