ui changes
1. Pathways changes
a. Move fear of death from Live your best life to manage your emotions - for adults and teens
b. Rename - Understand yourself in Pathway to - Learn to question yourself (can we do this for adults as well?)
c. Rename live your best life to - Succeed in life (for adults as well)
2. search image missing for social media and kindness